refactor(editor): minimize package store data API
